+++ crates/wasm_lite_std/run-browser-tests.sh
@@ -12,6 +12,6 @@ RUSTFLAGS="-C target-feature=+atomics,+bulk-memory,+mutable-globals \
  -C link-arg=--shared-memory -C link-arg=--max-memory=1073741824 -C link-arg=--import-memory \
  -C link-arg=--export=__stack_pointer -C link-arg=--export=__tls_base -C link-arg=--export=__tls_size \
  -C link-arg=--export=__tls_align -C link-arg=--export=__wasm_init_tls" \
-CARGO_TARGET_WASM32_UNKNOWN_UNKNOWN_RUNNER="$PWD/target/debug/wasm-lite run" \
+CARGO_TARGET_WASM32_UNKNOWN_UNKNOWN_RUNNER="$PWD/target/debug/wasm_lite run" \
 cargo +nightly test -p wasm_lite_std --test browser \
   --target wasm32-unknown-unknown -Z build-std=std,panic_abort "$@"